+++ tests/handlers/login_oidc_ldap_test.py
@@ -371,3 +371,36 @@ async def test_missing_attrs(
     assert r.status_code == 200
     assert r.headers["X-Auth-Request-User"] == "ldap-user"
     assert "X-Auth-Request-Email" not in r.headers
+
+
+@pytest.mark.asyncio
+async def test_invalidate_cache(
+    tmp_path: Path,
+    client: AsyncClient,
+    respx_mock: respx.Router,
+    mock_ldap: MockLDAP,
+) -> None:
+    config = await reconfigure(tmp_path, "oidc-ldap")
+    assert config.ldap
+    assert config.ldap.user_base_dn
+    mock_ldap.add_entries_for_test(
+        config.ldap.user_base_dn,
+        config.ldap.user_search_attr,
+        "ldap-user",
+        [{"uidNumber": ["2000"]}],
+    )
+    token = create_upstream_oidc_jwt(uid="ldap-user")
+
+    r = await simulate_oidc_login(client, respx_mock, token)
+    assert r.status_code == 403
+
+    # Add the group entry.  This should immediately succeed because the cache
+    # is invalidated.  If the cache were not invalidated, it would fail again.
+    mock_ldap.add_entries_for_test(
+        config.ldap.group_base_dn,
+        "member",
+        "ldap-user",
+        [{"cn": ["foo"], "gidNumber": ["1222"]}],
+    )
+    r = await simulate_oidc_login(client, respx_mock, token)
+    assert r.status_code == 307